+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilteredTableModel.java
@@ -434,7 +434,7 @@ final class DefaultFilteredTableModel<R, C> extends AbstractTableModel implement
 		});
 		addTableModelListener(removeSelectionListener);
 		filterModel.changeObserver().addListener(this::filterItems);
-		sortModel.addSortingChangedListener(columnIdentifier -> sortItems());
+		sortModel.sortingChangedObserver().addListener(this::sortItems);
 	}
 
 	private List<Object> columnValues(Stream<Integer> rowIndexStream, int columnModelIndex) {