+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilteredTableModelTest.java
@@ -467,7 +467,7 @@ public final class DefaultFilteredTableModelTest {
 	void sorting() {
 		AtomicInteger actionsPerformed = new AtomicInteger();
 		Consumer<Integer> listener = columnIdentifier -> actionsPerformed.incrementAndGet();
-		tableModel.sortModel().addSortingChangedListener(listener);
+		tableModel.sortModel().sortingChangedObserver().addDataListener(listener);
 
 		tableModel.refresh();
 		FilteredTableSortModel<TestRow, Integer> sortModel = tableModel.sortModel();
@@ -503,7 +503,7 @@ public final class DefaultFilteredTableModelTest {
 		sortModel.setSortOrder(0, SortOrder.DESCENDING);
 		assertEquals(tableModel.getRowCount() - 2, tableModel.indexOf(NULL));
 		sortModel.setSortOrder(0, SortOrder.UNSORTED);
-		tableModel.sortModel().removeSortingChangedListener(listener);
+		tableModel.sortModel().sortingChangedObserver().removeDataListener(listener);
 	}
 
 	@Test